Add *.iml to gitignore
We originally wanted to commit *.iml files for so that the sdk could
be opened in android studio just by doing File->Open. However, iml files
change so often that we decided it would be better to remove them from
version control. Now you will need to import the sdk into android studio
instead of just opening it.